Rename testfs.rs to fuse.rs to avoid name conflict with the module
