+++ tests/common/testfs/utils.rs
@@ -3,7 +3,7 @@
 //  * For the full copyright and license information, please view the LICENSE
 //  * file that was distributed with this source code.
 
-use crate::common::testfs::testfs::TestFs;
+use crate::common::testfs::fuse::TestFs;
 use fuser::BackgroundSession;
 use fuser::MountOption::FSName;
 use std::ffi::CString;
